requirements.txt
aiogram==3.*
python-dotenv==1.*
aiohttp==3.*
aiosqlite
google-auth-oauthlib
google-auth
google-api-python-client
python-dotenv